Polish (#2)
* Fix permitAll matcher

When a request to "/" is requested, the path that is presented to Spring
Security is "/index.jsp". With the current configuration that means that
"/index.jsp" will be compared against the pattern "/**/*.{js,html,css}" to
determine if the URL should be permitAll. The pattern matches because
"{js,html,css}" is a path variable name of "js,html,css" that can have
any value. In this case the value of the path variable name is
"js,html,css" with a value of "jsp".

This commit changes to using a regular expression to match on the file
suffix to determine if permitAll should be applied.

* Add SecurityConfigurationTests

* Use Beans

* Polish

* Shorten Property Names